requirements_pretraining.txt
# Core dependencies
torch>=2.0.0
numpy>=1.21.0
transformers>=4.21.0
datasets>=2.0.0
tokenizers>=0.13.0

# Data processing
PyYAML>=6.0
h5py>=3.7.0
pandas>=1.5.0

# Training and logging
wandb>=0.13.0
tqdm>=4.64.0
tensorboard>=2.10.0

# Optimization
accelerate>=0.21.0
deepspeed>=0.9.0  # Optional for large-scale training

# Development and testing
pytest>=7.0.0
black>=22.0.0
flake8>=5.0.0
mypy>=0.991

# Scientific computing
scipy>=1.9.0
scikit-learn>=1.1.0

# Genomics-specific (optional)
biopython>=1.79  # For FASTA parsing
pysam>=0.19.0    # For genomic file formats

# Visualization (optional)
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.10.0
